pyproject.toml
[tool.poetry]
name = "mlops-project"
version = "0.1.0"
description = "Predicting Likelihood of Lung Cancer, being the main goal to explore MLOps tools"
authors = ["MAGomes95 <alexandre95mario@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
pandas = "^2.1.1"
numpy = "^1.26.1"
seaborn = "^0.13.0"
matplotlib = "^3.8.0"
scikit-learn = "^1.3.2"
jupyterlab = "^4.0.7"
scipy = "^1.11.3"
fire = "^0.5.0"
optuna = "^3.4.0"
pydantic = "^2.4.2"
comet-ml = "^3.35.1"
python-dotenv = "^1.0.0"
cerebrium = "^1.13.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

